refactor: FontReadyContainer / MsdfText を削除し FontProvider に一本化
FontProvider が fontFamilies 付き Container + fontReady 遅延を内包するため、
配下では uikit の Text をそのまま使えるようになった。
FontReadyContainer と MsdfText は不要になったため削除。
