+++ src/fonts/FontProvider.tsx
@@ -1,4 +1,5 @@
-import { memo, useMemo, type ReactNode } from 'react'
+import { memo, useMemo, useState, useEffect, type ReactNode } from 'react'
+import { Container, Text } from '@react-three/uikit'
 import { FontContext } from './FontContext'
 import { useMsdfFont } from '../hooks/useMsdfFont'
 
@@ -13,10 +14,10 @@ interface Props {
 }
 
 /**
- * ツリー全体で1つの MSDF アトラスを共有するための Provider。
+ * ツリー全体で MSDF アトラスを共有し、配下で日本語テキストを直接使えるようにする Provider。
  *
  * `texts` に含まれる文字から MSDF フォントを1回だけ生成し、
- * 配下の `FontReadyContainer` が自動的にこのアトラスを利用する。
+ * `fontFamilies` を Container 経由で配下に適用する。
  * Suspense 対応のため、親に `<Suspense>` が必要。
  *
  * @example
@@ -25,8 +26,7 @@ interface Props {
  *
  * <Suspense fallback={null}>
  *   <FontProvider texts={ALL_TEXTS}>
- *     <VideoPlayer id="v1" />
- *     <LiveVideoPlayer id="l1" />
+ *     <Text fontSize={48}>URLを入力してください</Text>
  *   </FontProvider>
  * </Suspense>
  * ```
@@ -36,8 +36,23 @@ export const FontProvider = memo(
     const fontFamilies = useMsdfFont(texts, fontUrl, textureSize)
     const value = useMemo(() => ({ fontFamilies }), [fontFamilies])
 
+    // uikit はフォントテクスチャを非同期でロードするため、
+    // 初回レンダリングでデフォルトの inter フォントへフォールバックし
+    // "Missing glyph info" 警告が出る。スペースのみの Text で
+    // フォントロードをトリガーし、次フレームで children を表示する。
+    const [fontReady, setFontReady] = useState(false)
+    useEffect(() => {
+      setFontReady(false)
+      const id = requestAnimationFrame(() => setFontReady(true))
+      return () => cancelAnimationFrame(id)
+    }, [fontFamilies])
+
     return (
-      <FontContext.Provider value={value}>{children}</FontContext.Provider>
+      <FontContext.Provider value={value}>
+        <Container fontFamilies={fontFamilies}>
+          {fontReady ? children : <Text fontSize={1}>{' '}</Text>}
+        </Container>
+      </FontContext.Provider>
     )
   },
 )
